[#843] Make migrations independent of schema logic
Schema logic may change from the time when they were used in the
migration. Since migrations expect their targets to be in the place they
were when the migration was written and not in the place where the
schema logic thinks it is today, it could cause painful errors later
that this commit helps avoid.