Add multiwallet configuration to config file
This commit sets up the config file for use in the multiwallet refactor.
Each coin has its own config added with the idea that, abset the config, the
defaults will be used.

This commit does not change the existing wallet config object which will be
deprecated when the full refactor is complete.