Return cached profile if not found on network
Currently the profile fetching logic will try to query IPNS first if the cached
record is older than two weeks. The problem is if the IPNS query fails it does
not return the cached value and instead errors.

This commit changes it to return the cached value if the IPNS query fails.