Switch to modifying ttl field in the IPNS record when saving IPNS cache
Currently when usecache is selected we check the validity on the IPNS record and
if it less than two weeks old we return the cache. Otherwise we fetch from the
DHT. This implies that when we get a record back we update the validity field
with the new expiration (two weeks ahead) and resave in the database.

The problem is the validity field is covered by the IPNS singatures and updating
the validity breaks the signature. This is not a problem for normal nodes, but
gateways are currently serving signed IPNS records and, because of the validity
update, are returning bad signatures.

To rectify this we will switch from modifying the validity field, which is covered
by the signature, to the ttl field which is not.

Since we are saving this records for our own usage and not impacting IPNS, modifying
the ttl field will not have any negative affects.